requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.12" and python_version < "4.0"
bcrypt==4.3.0 ; python_version >= "3.12" and python_version < "4.0"
beautifulsoup4==4.13.4 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.7.14 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.2.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
deprecation==2.1.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.115.14 ; python_version >= "3.12" and python_version < "4.0"
gotrue==2.12.3 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.2.3 ; python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
h11==0.16.0 ; python_version >= "3.12" and python_version < "4.0"
h2==4.2.0 ; python_version >= "3.12" and python_version < "4.0"
hpack==4.1.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.12" and python_version < "4.0"
httpx[http2]==0.28.1 ; python_version >= "3.12" and python_version < "4.0"
hyperframe==6.1.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.12" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.12" and python_version < "4.0"
postgrest==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.11.7 ; python_version >= "3.12" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
realtime==2.6.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.4 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
soupsieve==2.7 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.41 ; python_version >= "3.12" and python_version < "4.0"
sqlmodel==0.0.24 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.46.2 ; python_version >= "3.12" and python_version < "4.0"
storage3==0.12.0 ; python_version >= "3.12" and python_version < "4.0"
strenum==0.4.15 ; python_version >= "3.12" and python_version < "4.0"
supabase==2.17.0 ; python_version >= "3.12" and python_version < "4.0"
supafunc==0.10.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.14.1 ; python_version >= "3.12" and python_version < "4.0"
typing-inspection==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.35.0 ; python_version >= "3.12" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.12" and python_version < "4.0"
